requirements.txt
# Google Cloud Kütüphaneleri
google-cloud-discoveryengine==0.11.16
google-cloud-storage==2.14.0
google-cloud-sql-connector==0.11.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Web Scraping ve Data Processing
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
scrapy==2.11.0

# Data Analysis ve ML
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Utility
python-dotenv==1.0.0
tqdm==4.66.1
pyyaml==6.0.1
lxml==4.9.3
urllib3==2.0.7 